pyproject.toml
[tool.poetry]
name = "irodolly"
version = "0.1.0"
description = "funcy analytics features for painting"
authors = ["snowsheep <71830342+sheep-snow@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.110.0"
boto3 = "^1.29.3"
opencv-python = "^4.8.1.78"
mediapipe = "^0.10.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
